+++ src/meltano/core/project_add_custom_service.py
@@ -23,8 +23,8 @@ class ProjectAddCustomService(ProjectAddService):
         )
         click.echo("- identifier to find related/compatible plugins")
         if plugin_type == PluginType.EXTRACTORS:
-            click.echo("- default value for the `schema` setting when used")
-            click.echo("  with loader 'target-postgres' or 'target-snowflake'")
+            click.echo("- preferred database schema (`preferred_schema` extra),")
+            click.echo("  for use by loaders that support a target schema")
         elif plugin_type == PluginType.LOADERS:
             click.echo("- default value for the `target` setting when used")
             click.echo("  with transformer 'dbt'")